【autopack测试脚本】 自动打包测试脚本

+++ test/util/autopack/start.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
+if [ -d "./result" ]; then
 rm -r result
+fi
 wget -O a.txt  "http://solar.baidu.com/autopack?return=json&fid=baidunuomi&svn=https://svn.baidu.com/app/search/tuangou/branches/mobile-bainuo/common/tuangou_4-3-102_BRANCH"
 wget -O b.txt  "http://fedev.baidu.com:8803/autopack?return=json&fid=baidunuomi&svn=https://svn.baidu.com/app/search/tuangou/branches/mobile-bainuo/common/tuangou_4-3-102_BRANCH"
 if [ ! -d "./result" ]; then
